add code

+++ util-zex/src/main/java/com/zifang/util/zex/date/DateUtils.java
@@ -593,7 +593,7 @@ public class DateUtils {
         }     
         n = n-1;     
         if(stype==2){     
-            n = (int)n/365;     
+            n = n /365;
         }        
         return n;     
     }
